+++ srcs/queue/new_qnode.c
@@ -6,14 +6,14 @@
 /*   By: rpinoit <rpinoit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/09 14:15:39 by rpinoit           #+#    #+#             */
-/*   Updated: 2019/02/09 14:19:33 by rpinoit          ###   ########.fr       */
+/*   Updated: 2019/04/01 23:59:30 by rpinoit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 #include "queue_42.h"
 
-t_qnode *new_qnode(int k)
+t_qnode *new_qnode(size_t k)
 {
     t_qnode *new;
